backend/requirements.txt
absl-py==1.0.0
astunparse==1.6.3
autofaiss==2.4.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
dataclasses==0.6
faiss-cpu==1.7.1.post3
fire==0.4.0
Flask==2.0.2
flatbuffers==2.0
fsspec==2021.11.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
numpy==1.21.5
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.3.5
Pillow==8.4.0
protobuf==3.19.1
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
tqdm==4.62.3
typing-extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
torch>=1.1
torchvision==0.5.0
sklearn
yacs
